+++ compiler-utils/src/testFixtures/java/com/squareup/anvil/compiler/internal/testing/AnvilCompilationMode.kt
@@ -13,5 +13,6 @@ public sealed class AnvilCompilationMode(public val analysisBackend: AnalysisBac
   ) : AnvilCompilationMode(EMBEDDED)
   public data class Ksp(
     val symbolProcessorProviders: List<SymbolProcessorProvider> = emptyList(),
+    val options: Map<String, String> = emptyMap(),
   ) : AnvilCompilationMode(KSP)
 }